1.1 Several changes
Updated to the 1.1.
Several visual and code changes.

# UPDATES:
Now, the app show all the ingredients in another ConstraintLayout.
The UI now is more beautiful and easy to work.
The strings.xml has been updated with more strings, making more easily to translate the app.

# KNOWN BUGS:
If you press the calculate button without any valor in the input, the app crashes. This will be fixed in future versions.

# COMING:
Change the "hint" of the grams input for another more confortable.
Small visual changes.
Bug fixes.