2.0 - Several changes
In this update, we added a configuration page. It's not changing the intern values, but saves data. Also, the app saves data.

# UPDATES:
Now, when you press the "CALCULATE" button, the value is saved into your data and when you open the app, this value will appear.
Also, there's added a configuration page for changing the values of the intern variables. For now, it only saves the values, but doesn't update the intern variables, so it's not doing anything (only visual).

# KNOWN BUGS:
When you press the calculate button without any value, the app crashes. But, at least, now the app will have always a value when you open it.

# COMING:
In next updates, there will be more visual updates. The config button will work, and will solve some bugs.