2.1 - Config page updated
In this update, the config page has been upgrated to a new level.

# UPDATES:
The config page now is working. When you put here values and press the save button, those valors are going to be the values that app uses.

# KNOWN BUGS:
Visual bug, the config page isn't aligned correctly.
When you don't put any value in an input, the app will crash.
The config page is saved, but the values aren't being imported to the intern variables until the save button is pressed.
The kewboard isn't dissapearing.