2.2 - Several changes
This is a very important update. The config page is working correctly, and the save button has been fixed. Now, it saves the values correctly.

# UPDATES:
The config page has been updated. The layout is displaying correctly.
And the values of the config page now are more easy to use. You don't have to put 0.2 for have a 20%, now you only have to put 20! The app will convert it.
When you press the save button, an alert (Toast) is displaying. That means all is working fine.
When you open the app, this will calculate all in that moment with the values it has. You don't need to press the calculate button when you open the app if the values are all correct, it will calculate all.

# BUG FIXED:
The layout of the config page is displaying correctly, all the items aligned.
The data is saving perfect.
When you press the save button in the config page, now those values are applying correctly to the main page, so the app is working fine. You don't have to push the save button twice, it's solved.
When you change of activity (go to Main or ConfigPage), now the app close the previous window. This is saving a lot of resources, so made the app works more fast.

# KNOWN BUGS:
The app crashes when you don't put a value in a input. This will be solved in a few updates.
The app isn't woring fine when you turn your phone. I'll update this as sooner I can.