2.4 - Visual issues solved
This is a simple update that solve a small bugs that happened when the font-size of the phone is not the default. There are some bugs that cannot be solved, at least from now.

# UPDATES:
Now, in the config page, there's a tant-per-cent character (%).

# BUG SOLVED:
When the font-size of a phone isn't the default (this can be changed in the settings of the phone), in the landscape mode, now the Main Page is working fine.
In the normal mode, the config page now is seeing correctly too.

# KNOWN BUGS:
The tant-per-cent, in the landscape mode, if the font-size of the phone isn't the default, isn't aligned with the number. This only happens in a very small cases.