2.5 - Augment and reduce buttons added
This update it isn't a big update.
It's small, but very necessary.

# UPDATES:
It has been added an augment and a reduce button to the quantity of bread you want. Now, if you press the + button, the quantity will augment 25, and if you press the - button, id will decrease 25 grams.

# COMING:
I'll try to put a dark mode into the app. It's not confirmed, but I'll try to put it.